Moved SettingsScreen under ui/screens/settings
